+++ setupTest.ts
@@ -1,7 +1,10 @@
 import { configure } from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
+
 import 'jest-styled-components'
+
 configure({ adapter: new Adapter() })
+
 
 /* eslint-disable @typescript-eslint/no-empty-function */
 window.IntersectionObserver = class {