+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Add initial SEO improvements](https://wealize.atlassian.net/browse/WLZAA-125)
+- [Add translations for blog](https://wealize.atlassian.net/browse/WLZAA-138)
 - [Add basic design to the html content that comes from Ghost](https://wealize.atlassian.net/browse/WLZAA-136)
 - [Add new type of PostCard to be able to link external content](https://wealize.atlassian.net/browse/WLZAA-134)
 - [Reduce top spacing in general containers](https://wealize.atlassian.net/browse/WLZAA-133)